+++ src/import_process/management/commands/import_spots.py
@@ -8,7 +8,7 @@ from django.core.management.base import BaseCommand
 from datasets.blackspots.models import Document, Spot
 from import_process.clean import clear_models
 from import_process.process_xls import process_xls
-from storage.objectstore import ObjectStore
+from storage.object_store import ObjectStore
 
 logging.basicConfig(level=logging.DEBUG)
 log = logging.getLogger(__name__)
