+++ src/main/settings.py
@@ -258,6 +258,7 @@ SWAGGER_SETTINGS = {
 }
 
 OBJECTSTORE_UPLOAD_CONTAINER_NAME = os.environ["OBJECTSTORE_UPLOAD_CONTAINER_NAME"]
+OBJECTSTORE_ENV = os.environ["OBJECTSTORE_ENV"]
 
 if DEBUG:
